# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/parsel/search": {
      "get": {
        "operationId": "ParselCompatController_search[0]",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/api/v1/parsel/search": {
      "get": {
        "operationId": "ParselCompatController_search[1]",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/parsel": {
      "get": {
        "operationId": "ParselCompatController_lookup[0]",
        "parameters": [
          {
            "name": "ada",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parsel",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "il",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ilce",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/api/v1/parsel": {
      "get": {
        "operationId": "ParselCompatController_lookup[1]",
        "parameters": [
          {
            "name": "ada",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parsel",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "il",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ilce",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/parsel/geometry/{id}": {
      "get": {
        "operationId": "ParselCompatController_geometry[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/api/v1/parsel/geometry/{id}": {
      "get": {
        "operationId": "ParselCompatController_geometry[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/parsel/{id}/context": {
      "get": {
        "operationId": "ParselCompatController_context[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_geometry",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/api/v1/parsel/{id}/context": {
      "get": {
        "operationId": "ParselCompatController_context[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_geometry",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/parsel/{id}/related-plans": {
      "get": {
        "operationId": "ParselCompatController_relatedPlans[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/api/v1/parsel/{id}/related-plans": {
      "get": {
        "operationId": "ParselCompatController_relatedPlans[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/parsel/{id}/summary": {
      "get": {
        "operationId": "ParselCompatController_summary[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
    "/api/v1/parsel/{id}/summary": {
      "get": {
        "operationId": "ParselCompatController_summary[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ParselCompat"
        ]
      }
    },
...
    "/watchlist/{id}": {
      "delete": {
        "operationId": "WatchlistCompatController_deleteWatchlist[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WatchlistCompat"
        ]
      }
    },
    "/api/v1/watchlist/{id}": {
      "delete": {
        "operationId": "WatchlistCompatController_deleteWatchlist[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "WatchlistCompat"
        ]
      }
    },
...
    "/reports/{id}": {
      "get": {
        "operationId": "ReportsCompatController_getReport[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ReportsCompat"
        ]
      }
    },
    "/api/v1/reports/{id}": {
      "get": {
        "operationId": "ReportsCompatController_getReport[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ReportsCompat"
        ]
      }
    },
...
  },
  "info": {
    "title": "Türkiye E-İmar Platform API",
    "description": "Backend foundation for source discovery, connector orchestration, parcel queries, and GIS analysis.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}